Moved types and also handled the apperance of widget trends
