+++ src/admin/hooks/customer-analytics.tsx
@@ -2,28 +2,7 @@ import { DateRange } from 'react-day-picker';
 import { useQuery } from '@tanstack/react-query';
 import type { UseQueryOptions } from '@tanstack/react-query';
 import { retrieveCustomersAnalytics } from '../lib/data/customer-analytics';
-
-export type CustomerAnalyticsResponse = {
-  total_customers: number;
-  new_customers: number;
-  returning_customers: number;
-  customer_count: {
-    name: string;
-    returning_customers: number;
-    new_customers: number;
-  }[];
-  customer_group: { name: string; total: number }[];
-  customer_sales: {
-    customer_id: string;
-    sales: number;
-    name: string;
-    groups: string[];
-    order_count: number;
-    last_order: Date;
-    email: string;
-  }[];
-  currency_code: string;
-};
+import { CustomerAnalyticsResponse } from '../../api/admin/agilo-analytics/customers/route';
 
 export const useCustomerAnalytics = (
   query: DateRange | undefined,
