+++ src/admin/hooks/product-analytics.tsx
@@ -2,20 +2,7 @@ import { DateRange } from 'react-day-picker';
 import { useQuery } from '@tanstack/react-query';
 import type { UseQueryOptions } from '@tanstack/react-query';
 import { retrieveProductAnalytics } from '../lib/data/product-analytics';
-
-export type ProductAnalyticsResponse = {
-  lowStockVariants: {
-    sku: string;
-    inventoryQuantity: number;
-    variantName: string;
-    variantId: string;
-    productId: string;
-  }[];
-  variantQuantitySold: {
-    title: string;
-    quantity: number;
-  }[];
-};
+import { ProductAnalyticsResponse } from '../../api/admin/agilo-analytics/products/route';
 
 export const useProductAnalytics = (
   query: DateRange | undefined,
